+++ Sources/Extensions/UIKit/UIViewController/UIViewController+PresentFullScreen.swift
@@ -1,15 +1,18 @@
 import UIkit
 
 public extension UIViewController {
-    /// Method present new ViewController modally in full screen mode
+
+    /// Method present new ViewController modally in full screen mode by default
     ///
     /// - Parameters:
     ///   - viewController: ViewController to present
-    ///   - animated: ppresent with animation
-    ///   - completion: completion block after present ended
-    func presentFullScreen(_ viewController: UIViewController,
-                           animated: Bool = true,
-                           completion: (() -> Void)? = nil) {
+    ///   - presentationStyle: presentation style (default is `fullScreen`
+    ///   - animated: present with animation or not
+    ///   - completion: completion block after present finish
+    func presentModal(_ viewController: UIViewController,
+                      presentationStyle: UIModalPresentationStyle = .fullScreen,
+                      animated: Bool = true,
+                      completion: (() -> Void)? = nil) {
 
         viewController.modalPresentationStyle = .fullScreen
         present(viewController, animated: animated, completion: completion)